Move form to separated file
